Refactor governor module: move GovernorContract and types to internal/contracts, update imports across the application
