+++ ai-agent/src/internal/contracts/index.ts
@@ -1,0 +1,2 @@
+export * from "./governor";
+export * from "./types";
