+++ ai-agent/src/internal/contracts/types.ts
