--- ai-agent/src/internal/governor/index.ts
@@ -1,2 +1,0 @@
-export * from "./contract";
-export * from "./types";
